LaserJet PostScript 4000 PPD bugfix
The PostScript PPD for the HP LaserJet 4000 is buggy, and prevent printing complex documents.
The 4000 and 4050 are pretty much the same printer. The fix was to extract the missing snipset from the PPD for the 4050 and put it in the PPD for the 4000.

Origin: vendor
Bug: https://bugs.launchpad.net/hplip/+bug/1298194
Bug-Debian: https://bugs.debian.org/742766
Patch-Name: HP-LaserJet_4000-PostScript-PPD.patch

Gbp-Pq: Name HP-LaserJet_4000-PostScript-PPD.patch